// resources/js/types/theme.ts
export interface ThemeColors {
    backgroundColor: string;
    backgroundAltColor: string;
    panelColor: string;
    lineColor: string;
    lineLightColor: string;
    inkColor: string;
    inkMutedColor: string;
    inkDimColor: string;
    successColor: string;
    accentColor: string;
    highlightColor: string;
    warningColor: string;
    orangeColor: string;
    dangerColor: string;
    primaryColor: string;
}